Fix cumulative values not being passed to next day.
When one day is missing from the spreadsheet, we assume that day's cumulative number is 0.

Instead the daily summary now tries to pass one the cumulative number to the next day if that number is missing (0).

Also adds a verification step that will throw and error if the data isn't valid.

Moves JSON writing to generate.js to reduce the amount of concerns fetch* and summarize* need to care about.