Replace drive-db with custom Google Sheets fetcher
drive-db uses the Sheets v3 API that is deprecated and also fragile to tab order.

fetch_sheet.js is a more simple fetcher that does the same thing using the v4 API using tab names instead of tab order.

Update the rest of the code to use this fetcher instead.