Charts axis uses absolute max value as label rather than average.
When comparing the chart to the number at the top, the average max value can be way different than the absolute value. But it is confusing if the value is not similar, so instead, we use the absolute max value as the label but keep the chart using 7-day active.