Allow modification of the confirmed numbers rather than just the recovery numbers.
This more accurately models the discrepencies we're seeing.

Apply the correction over time by smearing the difference across days (for the daily summary only) so that it doesn't appear as a big spike at the end.

Temporarily not apply the confirmed correction until we update the data in one go.