Huge optimization for the patient counts.
Using patientCount rather than blowing up the array of patient data.

We used to output one row for every patient when internally summing the numbers. But this new wave has caused this to consume way too much memory and kill the sorting algorithm.

So we're no longer doing this so that our array length doesn't increase exponentionally